Create BMI Calculator
In this exercise, I created a very simple Body Mass Index (BMI) calculator to know if it is urgent for you to visit a doctor and good nutritionist, or you are rocking life like a pro.  

So, for those who are wondering what BMI is used for, here is a quick definition. Basically, BMI is a measurement of a person's leanness or corpulence based on their height and weight, and is intended to quantify tissue mass. It is widely used as a general indicator of whether a person has a healthy body weight for their height or not. 

Have a play around and have fun.