Create FizzBuzz
Pretty tricky child's game, are you up for it? 

Instructions:
We are going to go through a range of 1 to a 100.
When the number is divisible by 3 then instead of printing the number it should print "Fizz".
When the number is divisible by 5, then instead of printing the number it should print "Buzz".`
And if the number is divisible by both 3 and 5 e.g. 15 then instead of the number it should print "FizzBuzz"

Let's give it a go!